Extract AlignmentErrors
