+++ render-ws-java-client/src/main/java/org/janelia/render/client/newsolver/StackAlignmentComparisonClient.java
@@ -13,7 +13,6 @@ import org.janelia.alignment.match.OrderedCanvasIdPairWithValue;
 import org.janelia.alignment.spec.Bounds;
 import org.janelia.alignment.spec.ResolvedTileSpecsWithMatchPairs;
 import org.janelia.alignment.spec.TileSpec;
-import org.janelia.alignment.util.FileUtil;
 import org.janelia.render.client.ClientRunner;
 import org.janelia.render.client.RenderDataClient;
 import org.janelia.render.client.newsolver.solvers.WorkerTools;
@@ -25,12 +24,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 import java.util.function.DoubleBinaryOperator;
-import java.util.stream.Collectors;
 
 public class StackAlignmentComparisonClient {
 
@@ -161,7 +157,6 @@ public class StackAlignmentComparisonClient {
 		return alignmentErrors;
 	}
 
-
 	public enum DifferenceMetric {
 		RELATIVE((a, b) -> Math.abs(a - b) / a),
 		ABSOLUTE((a, b) -> Math.abs(a - b));
@@ -170,43 +165,6 @@ public class StackAlignmentComparisonClient {
 
 		DifferenceMetric(final DoubleBinaryOperator metricFunction) {
 			this.metricFunction = metricFunction;
-		}
-	}
-
-	private static class AlignmentErrors {
-		private final List<OrderedCanvasIdPairWithValue> pairwiseErrors = new ArrayList<>();
-
-		public void addError(final OrderedCanvasIdPair pair, final double error) {
-			pairwiseErrors.add(new OrderedCanvasIdPairWithValue(pair, error));
-		}
-
-		public void absorb(final AlignmentErrors other) {
-			pairwiseErrors.addAll(other.pairwiseErrors);
-		}
-
-		public List<OrderedCanvasIdPairWithValue> getWorstPairs(final int n) {
-			return pairwiseErrors.stream()
-					.sorted((p1, p2) -> Double.compare(p2.getValue(), p1.getValue()))
-					.limit(n)
-					.collect(Collectors.toList());
-		}
-
-		public static AlignmentErrors computeDifferences(final AlignmentErrors baseline, final AlignmentErrors other, final DoubleBinaryOperator comparisonMetric) {
-			final AlignmentErrors differences = new AlignmentErrors();
-			final Map<OrderedCanvasIdPair, Double> errorLookup = other.pairwiseErrors.stream()
-					.collect(Collectors.toMap(OrderedCanvasIdPairWithValue::getPair, OrderedCanvasIdPairWithValue::getValue));
-
-			baseline.pairwiseErrors.forEach(pairWithValue -> {
-				final double otherError = errorLookup.get(pairWithValue.getPair());
-				final double errorDifference = comparisonMetric.applyAsDouble(pairWithValue.getValue(), otherError);
-				differences.addError(pairWithValue.getPair(), errorDifference);
-			});
-
-			return differences;
-		}
-
-		public static void writeToFile(final AlignmentErrors errors, final String filename) throws IOException {
-			FileUtil.saveJsonFile(filename, errors.pairwiseErrors);
 		}
 	}
